chore: rename API
