--- packages/iceworks-server/src/app/io/controller/home/help.ts
@@ -1,11 +1,0 @@
-import * as path from 'path';
-
-export default (app) => {
-  const { Controller } = app;
-
-  return class HelpController extends Controller {
-    async getPath({ args }) {
-      return path.join(...args);
-    }
-  };
-};
