+++ packages/iceworks-server/src/app/router.ts
@@ -13,8 +13,6 @@ export default (app: Application) => {
     ['project.index.add', project.index.add],
     ['project.index.current', project.index.getCurrent],
     ['project.index.setCurrent', project.index.setCurrent],
-    ['project.index.openFolder', project.index.openFolder],
-    ['project.index.openEditor', project.index.openEditor],
     ['project.page.list', project.page.list],
     ['project.page.delete', project.page.delete],
     ['project.page.create', project.page.create],
@@ -30,9 +28,11 @@ export default (app: Application) => {
     ['material.index.resource', material.index.resource],
     ['material.index.getOne', material.index.getOne],
     ['material.index.recommendScaffolds', material.index.getRecommendScaffolds],
-    ['home.help.getPath', home.help.getPath],
-    ['home.index.workFolder', home.index.getWorkFolder],
-    ['home.index.setWorkFolder', home.index.setWorkFolder],
+    ['home.setting.workFolder', home.setting.getWorkFolder],
+    ['home.setting.setWorkFolder', home.setting.setWorkFolder],
+    ['home.system.getPath', home.system.getPath],
+    ['home.system.openFolder', home.system.openFolder],
+    ['home.system.openEditor', home.system.openEditor],
   ];
 
   routers.forEach(([eventName, handle]) => {
